script/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pymongo==4.10.1
tqdm==4.66.6